pyproject.toml
[project]
name = "swagdigitclassifier3000"
version = "0.1.0"
description = ""
authors = [
    {name = "PhMaly",email = "philemon.maly@le-campus-nemrique.fr"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "numpy (>=2.2.3,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "jupyterlab (>=4.3.5,<5.0.0)",
    "jupyter (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "swagdigitclassifier3000", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
